fix(sui-widget-embedder): more resilient removing of plugin and reuse functionality
